LTOBD2/LTOBD2PID_MONITOR_STATUS_THIS_DRIVE_CYCLE_41: fix potential crash by adding "dummy" results
This fixes a potential crash for applications showing a monitor status grid based on the number
of monitor results achieved by querying LTOBD2PID_MONITOR_STATUS_SINCE_DTC_CLEARED_01 and
LTOBD2PID_MONITOR_STATUS_THIS_DRIVE_CYCLE_41. Sane vehicles will report
a) the same number of monitor results for both PIDs, and
b) report the same type of ignition (which affects the number of available monitoring results) for both PIDs.

I have got a report where 0101 = 00 06 80 00 and 0141 = 00 4F FF 00.
This is completely nuts as it reports spark ignotion for 0101 and compression ignition for 0141,
which leads to a crash because the number of monitor results is inconsistent.

I'm working around this for now by adding two dummy monitor results to the compression ignition part.
This does not make the results consistent, but at least avoids the crash until we find out whether
this really is a bug in the vehicle's OBD2 implementation or whether we overlooked something in
the way we interpret the PID results.