automotive: LTOBD2PID: fix parsing the response to PID 03 (DTC)
Funny Story: I have developed this library against a commercially available
OBD2 simulator (will omit the brand to protect the guilty...), which worked
fine so far. For some reason though it contains a broken implementation of the
response to PID 03, which is supposed to be 43 [DTC1HI] [DTC1LO] ... This simulator
inserts the actual number of DTC elements before the DTC bytes, which --
according to SAE J1979 -- is wrong. Closes #20, thanks to JRJian (and others) for reporting!