LTOBD2ProtocolISO15764_4: swallow negative responses with 0x78
As per ISO 14229-1, a negative response with type RequestCorrectlyReceivedResponsePending is just an
intermediate response which is to be expected whenever the ECU will perform an operation
that could take "a bit longer". For OBD2 this should never happen – for UDS, it seems to be quite common.
In this case, we assume that another – positive – response is forthcoming, and just swallow the "negative" response.

NOTE: This is actually not a CANism, but independent from the underlying protocol,
hence it should also be implemented for the other protocols.